Split ButtonBar.ts
